fix(chat): surface llamacpp backend errors during a plain token stream
hasActiveLlamacppRequest gated only on abortControllers/busyThreads/
loadingModels/streamingContents — none of which are set during an ordinary
token stream, so a mid-generation ggml/backend crash was dropped before
setBackendError could fire. Add currentStreamThreadId (the reliable
in-flight signal) to the check. Provider separation is preserved: the gate
still early-returns when selectedProvider !== 'llamacpp', and the banner
masks on isLlamacppActive, so a router crash never decorates chats on other
providers. Extract the pure helpers to llamacppRouterError.ts (avoids the
react-refresh lint warning on the exported helper) and add tests.
